+++ .ccls-cache/@home@runner@Trabalho-Jogo-2D/src@class@ListaEntidades@ListaEntidades.cpp
@@ -5,9 +5,7 @@
 using namespace Entidades;
 using namespace Listas;
 
-ListaEntidades::ListaEntidades() {
-  this->graphicManager = GraphicManager::getInstance();
-}
+ListaEntidades::ListaEntidades() {}
 
 ListaEntidades::~ListaEntidades() {}
 
@@ -15,19 +13,6 @@ void ListaEntidades::push(Entidade *entidade) { list += entidade; }
 
 void ListaEntidades::executar() {
   for (int i = 0; i < list.getLength(); i++) {
-
-    if (i == 0) {
-      std::cout << "Before" << '\n';
-      std::cout << "x: " << list[i]->getSprite()->getPosition().x << '\n';
-      std::cout << "y: " << list[i]->getSprite()->getPosition().y << '\n';
-    }
-
     list[i]->executar();
-
-    if (i == 0) {
-      std::cout << "After" << '\n';
-      std::cout << "x: " << list[i]->getSprite()->getPosition().x << '\n';
-      std::cout << "y: " << list[i]->getSprite()->getPosition().y << '\n';
-    }
   }
 }