+++ CHANGELOG.md
@@ -5,5 +5,12 @@
 - Removido metodo "void imprimirse();" em class Ente
 - Removido atributo "int id;" em class Ente
 - Muda atributo "pos" para "position" em class Entidade
-- Adicionado atributo "velocity" em class Entidade
+- Adicionado atributo "flutuante" "collisionManager" em class Entidade
 - Adicionado metodos "draw" "move" "gravity" em class Entidade
+- Alterado muitas coisas em todas as classes herdadas de Entidade
+- Excluido "graphicManager" em class "ListaEntidades"
+- Adicionado metodo "virtual Entidade* clone();" em class "Entidade"
+- Alterado "void insertEntidade(Entidade* entidade);" em class "Fase"
+- Adicionado "Entidade* clone();" em class Jogador
+- Removido atributos de caverna e ruinas em class Jogo
+- Adicionado "Fase* fase;" em class Jogo